refactor(connect): move dynamic settings to connect settings
